Modest first contribution: MACsec Short Length bugfix :-)
Asking MACsec to protect (only, not encrypt, for clarity here) Ethernet frames that have only "A" as their payload, then "AA" and so on for example, produces the following in the network interface underneath the MACsec one, where packet number (pn) 1 for example has a short length (sl) specified as 3 due to the payload being only the original 0x88b6 Ethernet type followed by 0x41 for "A" in ASCII encoding; the remaining 8 bytes of each frame correspond to MACsec's Integrity Check Value here configured to its minimum length:

	05:58:03.988070 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 31: an 0, pn 1, flags CS, sl 3,
	        0x0000:  88b6 4199 3f51 a78d 6187 d8              ..A.?Q..a..

	05:58:05.023913 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 32: an 0, pn 2, flags CS, sl 4,
	        0x0000:  88b6 4141 ec86 7669 61b2 a9b8            ..AA..via...
[...]
	05:58:34.039732 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 59: an 0, pn 29, flags CS, sl 31,
	        0x0000:  88b6 4141 4141 4141 4141 4141 4141 4141  ..AAAAAAAAAAAAAA
	        0x0010:  4141 4141 4141 4141 4141 4141 4141 41be  AAAAAAAAAAAAAAA.
	        0x0020:  cf3e ae7a 35e8 2f                        .>.z5./

	05:58:35.068262 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 60: an 0, pn 30, flags CS, sl 32,
	        0x0000:  88b6 4141 4141 4141 4141 4141 4141 4141  ..AAAAAAAAAAAAAA
	        0x0010:  4141 4141 4141 4141 4141 4141 4141 4141  AAAAAAAAAAAAAAAA
	        0x0020:  f4c0 f652 851b 03b4                      ...R....

However, due to minimum frame sizes such as the >=64byte (including the 4 bytes of Frame Check Sequence that tcpdump does not see hence its "length 60" mentions above and below do meet this) requirement of Ethernet, such short frames are likely (depending on all the encapsulation possibilities) to get padded on their way to the receiver which in my case sees the above as:

	05:58:03.949439 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 60: an 0, pn 1, flags CS, sl 3,
	        0x0000:  88b6 4199 3f51 a78d 6187 d800 0000 0000  ..A.?Q..a.......
	        0x0010:  0000 0000 0000 0000 0000 0000 0000 0000  ................
	        0x0020:  0000 0000 0000 0000                      ........

	05:58:04.985249 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 60: an 0, pn 2, flags CS, sl 4,
	        0x0000:  88b6 4141 ec86 7669 61b2 a9b8 0000 0000  ..AA..via.......
	        0x0010:  0000 0000 0000 0000 0000 0000 0000 0000  ................
	        0x0020:  0000 0000 0000 0000                      ........
[...]
	05:58:34.000959 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 60: an 0, pn 29, flags CS, sl 31,
	        0x0000:  88b6 4141 4141 4141 4141 4141 4141 4141  ..AAAAAAAAAAAAAA
	        0x0010:  4141 4141 4141 4141 4141 4141 4141 41be  AAAAAAAAAAAAAAA.
	        0x0020:  cf3e ae7a 35e8 2f00                      .>.z5./.

	05:58:35.029493 08:00:27:c7:b8:1f > 08:00:27:69:71:09, ethertype 802.1AE MACsec (0x88e5), length 60: an 0, pn 30, flags CS, sl 32,
	        0x0000:  88b6 4141 4141 4141 4141 4141 4141 4141  ..AAAAAAAAAAAAAA
	        0x0010:  4141 4141 4141 4141 4141 4141 4141 4141  AAAAAAAAAAAAAAAA
	        0x0020:  f4c0 f652 851b 03b4                      ...R....

This is why for payloads shorter than 48 bytes (Ethernet's 64 minimum minus MACsec's header length of 8 when the optional Secure Channel Identifier is not included, and minus MACsec's ICV footer minimum length of 8) the sender has to fill in the Short Length value to help the receiver find the ICV since padding invalidates the "remaining 8 bytes of each frame" idea. Unfortunately Linux's MACsec implementation has a typo which makes it drop all padded short length frames: in my example above, pn 30 with thirty A's at 05:58:35 is the very first frame that makes it through because it is the first one big enough to not require padding.